backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "celery>=5.6.2",
    "fastapi>=0.129.2",
    "fastapi-mcp>=0.4.0",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.2.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "sqlmodel>=0.0.37",
    "uvicorn[standard]>=0.41.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pre-commit>=4.1.0",
    "pytest-mock>=3.14.0",
    "factory-boy>=3.3.3",
    "ipython>=9.0.0",
    "freezegun>=1.5.1",
    "faker>=36.0.0",
    "rich>=13.9.0",
]

test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.14.0",
    "factory-boy>=3.3.3",
    "freezegun>=1.5.1",
    "faker>=36.0.0",
]

[tool.ruff]
target-version = "py313"
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "C4",     # flake8-comprehensions
    "PT",     # flake8-pytest-style
]
ignore = [
    "E501",   # line too long (ruff fmt handles this)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["celery.*", "langchain.*", "langgraph.*", "slowapi.*"]
ignore_missing_imports = true
